DemoOne/src/app/auth.service.spec.ts
import { TestBed, async,ComponentFixture } from '@angular/core/testing';
import { RouterTestingModule } from '@angular/router/testing';
import {AngularFireModule} from "@angular/fire";
import { AngularFirestoreModule } from '@angular/fire/firestore';
import {AngularFireDatabaseModule} from '@angular/fire/database';
import {AngularFireAuthModule, AngularFireAuth} from "@angular/fire/auth";
import { ReactiveFormsModule } from '@angular/forms';
import { FirestoreSettingsToken} from '@angular/fire/firestore';
import { AuthService } from './auth.service';
import { LoginComponent } from './login/login.component';
let component: LoginComponent;
let fixture: ComponentFixture<LoginComponent>;

//This is for connecting to firebase and using the authentication and login services
var config = {
  apiKey: "AIzaSyBZBjbJQEstJ1Qjx9bP8kIejypgXUsN1CM",
  authDomain: "capstoneusers-b474f.firebaseapp.com",
  databaseURL: "https://capstoneusers-b474f.firebaseio.com",
  projectId: "capstoneusers-b474f",
  storageBucket: "capstoneusers-b474f.appspot.com",
  messagingSenderId: "754403435763"
};
//

describe('AuthService', () => {
  beforeEach(async(() => TestBed.configureTestingModule({
    imports: [
      RouterTestingModule,
      AngularFireModule.initializeApp(config),
      AngularFireAuthModule,
      ReactiveFormsModule,
      AngularFirestoreModule,
      AngularFireDatabaseModule
      
    ],
    declarations:[ LoginComponent],
    providers: [{ provide: FirestoreSettingsToken, useValue: {} },{provide:AngularFireAuthModule},{provide:AngularFireAuth},{provide:AuthService}],
    
  })));
  // beforeEach(()=>
  // {
  //   fixture = TestBed.createComponent(LoginComponent);
  //   component = fixture.componentInstance;
  //   fixture.detectChanges();
  // });

  it('Login check should be false', async(() => {
    expect(false).toBeFalsy();
  }));
  it('Login check should be True', async(() => {
    expect(false).toBeFalsy();
  }));
  it('Sending data should fail', async(() => {
    expect(false).toBeFalsy();
  }));
  it('Sending data should Succeed', async(() => {
    expect(false).toBeFalsy();
  }));
 
 
});
